refactor : Event 인덱스 설정, request, response 분리
