+++ src/main/java/com/spotlightspace/core/event/domain/Event.java
@@ -1,19 +1,10 @@
 package com.spotlightspace.core.event.domain;
 
 import com.spotlightspace.common.entity.Timestamped;
-import com.spotlightspace.core.event.dto.CreateEventRequestDto;
+import com.spotlightspace.core.event.dto.request.CreateEventRequestDto;
 import com.spotlightspace.core.user.domain.User;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
+
 import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -24,7 +15,11 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Getter
 @Entity
-@Table(name = "events")
+@Table(name = "events",
+        indexes = {
+                @Index(name = "idx_title_category", columnList = "title, category"),
+                @Index(name = "idx_location_category", columnList = "location, category")
+        })
 public class Event extends Timestamped {
 
     @Id
