+++ src/main/java/com/spotlightspace/core/event/repository/EventQueryRepository.java
@@ -1,7 +1,7 @@
 package com.spotlightspace.core.event.repository;
 
-import com.spotlightspace.core.event.dto.GetEventResponseDto;
-import com.spotlightspace.core.event.dto.SearchEventRequestDto;
+import com.spotlightspace.core.event.dto.response.GetEventResponseDto;
+import com.spotlightspace.core.event.dto.request.SearchEventRequestDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
