+++ src/main/java/com/spotlightspace/core/event/repository/EventQueryRepositoryImpl.java
@@ -4,10 +4,9 @@ import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.spotlightspace.core.event.domain.Event;
 import com.spotlightspace.core.event.domain.QEvent;
-import com.spotlightspace.core.event.dto.GetEventResponseDto;
-import com.spotlightspace.core.event.dto.SearchEventRequestDto;
+import com.spotlightspace.core.event.dto.response.GetEventResponseDto;
+import com.spotlightspace.core.event.dto.request.SearchEventRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
