+++ src/main/java/com/spotlightspace/core/event/service/EventService.java
@@ -5,7 +5,12 @@ import com.spotlightspace.common.entity.TableRole;
 import com.spotlightspace.common.exception.ApplicationException;
 import com.spotlightspace.core.attachment.service.AttachmentService;
 import com.spotlightspace.core.event.domain.Event;
-import com.spotlightspace.core.event.dto.*;
+import com.spotlightspace.core.event.dto.request.CreateEventRequestDto;
+import com.spotlightspace.core.event.dto.request.SearchEventRequestDto;
+import com.spotlightspace.core.event.dto.request.UpdateEventRequestDto;
+import com.spotlightspace.core.event.dto.response.CreateEventResponseDto;
+import com.spotlightspace.core.event.dto.response.GetEventResponseDto;
+import com.spotlightspace.core.event.dto.response.UpdateEventResponseDto;
 import com.spotlightspace.core.event.repository.EventRepository;
 import com.spotlightspace.core.eventticketstock.domain.EventTicketStock;
 import com.spotlightspace.core.ticket.repository.TicketRepository;
