+++ src/main/java/com/spotlightspace/core/point/dto/response/CreatePointResponseDto.java
@@ -1,4 +1,4 @@
-package com.spotlightspace.core.point.dto;
+package com.spotlightspace.core.point.dto.response;
 
 import com.spotlightspace.core.point.domain.Point;
 import lombok.AllArgsConstructor;
@@ -6,19 +6,19 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class GetPointResponseDto {
-
+public class CreatePointResponseDto {
     private Long id;
     private int amount;
     private String nickname;
-    
-    private GetPointResponseDto(Point point) {
+
+    private CreatePointResponseDto(Point point) {
         this.id = point.getId();
         this.amount = point.getAmount();
         this.nickname = point.getUser().getNickname();
     }
-    
-    public static GetPointResponseDto from(Point point) {
-        return new GetPointResponseDto(point);
+
+    public static CreatePointResponseDto from(Point point) {
+        return new CreatePointResponseDto(point);
     }
+
 }
