+++ src/main/java/com/spotlightspace/core/point/dto/response/GetPointResponseDto.java
@@ -1,4 +1,4 @@
-package com.spotlightspace.core.point.dto;
+package com.spotlightspace.core.point.dto.response;
 
 import com.spotlightspace.core.point.domain.Point;
 import lombok.AllArgsConstructor;
@@ -6,19 +6,19 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class CreatePointResponseDto {
+public class GetPointResponseDto {
+
     private Long id;
     private int amount;
     private String nickname;
-
-    private CreatePointResponseDto(Point point) {
+    
+    private GetPointResponseDto(Point point) {
         this.id = point.getId();
         this.amount = point.getAmount();
         this.nickname = point.getUser().getNickname();
     }
-
-    public static CreatePointResponseDto from(Point point) {
-        return new CreatePointResponseDto(point);
+    
+    public static GetPointResponseDto from(Point point) {
+        return new GetPointResponseDto(point);
     }
-
 }
