+++ src/test/java/com/spotlightspace/core/point/service/PointServiceTest.java
@@ -3,9 +3,9 @@ package com.spotlightspace.core.point.service;
 import com.spotlightspace.common.annotation.AuthUser;
 import com.spotlightspace.common.exception.ApplicationException;
 import com.spotlightspace.core.point.domain.Point;
-import com.spotlightspace.core.point.dto.CreatePointRequestDto;
-import com.spotlightspace.core.point.dto.CreatePointResponseDto;
-import com.spotlightspace.core.point.dto.GetPointResponseDto;
+import com.spotlightspace.core.point.dto.request.CreatePointRequestDto;
+import com.spotlightspace.core.point.dto.response.CreatePointResponseDto;
+import com.spotlightspace.core.point.dto.response.GetPointResponseDto;
 import com.spotlightspace.core.point.repository.PointRepository;
 import com.spotlightspace.core.user.domain.User;
 import com.spotlightspace.core.user.repository.UserRepository;
